Remove `plugin-api` module
