+++ plugin-base/build.gradle.kts
@@ -37,7 +37,6 @@ dependencies {
     compileOnlyApi(gradleApi())
     compileOnlyApi(Protobuf.GradlePlugin.lib)
     compileOnlyApi(Kotlin.GradlePlugin.api)
-    api(project(":plugin-api"))
     api(project(":tool-base"))
     compileOnly(gradleKotlinDsl())
     implementation(Logging.lib)
